+++ src/dialogs/CloneDialog.cpp
@@ -16,7 +16,6 @@
 #include "log/LogView.h"
 #include "ui/ExpandButton.h"
 #include "ui/RemoteCallbacks.h"
-#include "util/WaitCursor.h"
 #include <QApplication>
 #include <QCheckBox>
 #include <QComboBox>
@@ -301,7 +300,6 @@ public:
                                      "origin", mWatcher);
 
     entry->setBusy(true);
-    WaitCursor::track(mWatcher);
     mWatcher->setFuture(
         QtConcurrent::run(&git::Remote::clone, mCallbacks, url, path, bare));
   }