+++ src/ui/TabWidget.h
@@ -39,6 +39,8 @@ public:
 
   QIcon tabIcon(int index) const;
   void setTabIcon(int index, const QIcon &icon);
+  bool tabBusy(int index) const;
+  void setTabBusy(int index, bool busy);
   QString tabText(int index) const;
   void setTabText(int index, const QString &text);
   QString tabToolTip(int index) const;